Organize code
- Create util package for 2e2 tests
- Fix error for postgres healthcheck
- Use YAML anchors for docker-compose
- Add run_tests.sh script
