+++ test/events.e2e-spec.ts
@@ -4,30 +4,9 @@ import { AppModule } from '../src/app.module';
 import { EventsService } from '../src/routes/events/events.service';
 import { WebhookService } from '../src/routes/webhook/webhook.service';
 import { Webhook } from '../src/routes/webhook/entities/webhook.entity';
-import { connect as amqplibConnect, Connection } from 'amqplib';
 import { QueueProvider } from '../src/datasources/queue/queue.provider';
 import { TxServiceEventType } from '../src/routes/events/event.dto';
-
-async function publishMessage(
-  amqpUrl: string,
-  exchange: string,
-  queue: string,
-  msg: object,
-): Promise<boolean> {
-  const conn: Connection = await amqplibConnect(amqpUrl);
-  const channel = await conn.createChannel();
-  await channel.assertExchange(exchange, 'fanout', { durable: true });
-  // Make sure queue is binded to the exchange, as this function can be called before subscribing
-  await channel.assertQueue(queue, { durable: true });
-  const isMessagePublished = channel.publish(
-    exchange,
-    '',
-    Buffer.from(JSON.stringify(msg)),
-  );
-  await channel.close();
-  await conn.close();
-  return isMessagePublished;
-}
+import { publishMessage } from './util';
 
 describe('Events handling', () => {
   let app: INestApplication;
